core: Create config sub folder for bonder fee bps and multipler config
