+++ packages/core/src/addresses/types.ts
@@ -103,26 +103,7 @@ export type Bonders = {
   WBTC?: Routes
 }
 
-type Bps = {
-  ethereum: number
-  polygon: number
-  xdai: number
-  optimism: number
-  arbitrum: number
-}
-
-export type Fees = {
-  USDC?: Bps
-  USDT?: Bps
-  DAI?: Bps
-  MATIC?: Bps
-  ETH?: Bps
-  WBTC?: Bps
-}
-
 export type Addresses = {
   bridges: Bridges
   bonders: Bonders
-  bonderFeeBps?: Fees
-  destinationFeeGasPriceMultiplier?: number
 }
