+++ packages/core/src/config/goerli.ts
@@ -1,0 +1,15 @@
+import { Config } from './types'
+
+export const config : Config = {
+  bonderFeeBps: {
+    USDC: {
+      ethereum: 0,
+      polygon: 0,
+    },
+    DAI: {
+      ethereum: 0,
+      polygon: 0,
+    },
+  },
+  destinationFeeGasPriceMultiplier: 1
+}
