+++ packages/core/src/config/index.ts
@@ -1,0 +1,5 @@
+export * from './types'
+export { config as kovan } from './kovan'
+export { config as goerli } from './goerli'
+export { config as mainnet } from './mainnet'
+export { config as staging } from './staging'
