+++ packages/core/src/config/staging.ts
@@ -1,0 +1,14 @@
+import { Config } from './types'
+
+export const config : Config = {
+  bonderFeeBps: {
+    USDC: {
+      ethereum: 14,
+      polygon: 14,
+      xdai: 25,
+      optimism: 14,
+      arbitrum: 14
+    },
+  },
+  destinationFeeGasPriceMultiplier: 1.2
+}
