+++ packages/sdk/src/config/kovan.ts
@@ -1,6 +1,7 @@
 import { Chains } from './types'
 import { kovan as kovanAddresses } from '@hop-protocol/core/addresses'
 import { kovan as networks } from '@hop-protocol/core/networks'
+import { kovan as kovanConfig } from '@hop-protocol/core/config'
 
 const chains: Chains = {}
 
@@ -17,7 +18,7 @@ for (const chain in networks) {
 
 const addresses = kovanAddresses.bridges
 const bonders = kovanAddresses.bonders
-const bonderFeeBps = kovanAddresses.bonderFeeBps
-const destinationFeeGasPriceMultiplier = kovanAddresses.destinationFeeGasPriceMultiplier
+const bonderFeeBps = kovanConfig.bonderFeeBps
+const destinationFeeGasPriceMultiplier = kovanConfig.destinationFeeGasPriceMultiplier
 
 export { addresses, chains, bonders, bonderFeeBps, destinationFeeGasPriceMultiplier }
