+++ packages/sdk/src/config/staging.ts
@@ -1,6 +1,7 @@
 import { Chains } from './types'
 import { mainnet as networks } from '@hop-protocol/core/networks'
 import { staging as stagingAddresses } from '@hop-protocol/core/addresses'
+import { staging as stagingConfig } from '@hop-protocol/core/config'
 
 const chains: Chains = {}
 
@@ -17,7 +18,7 @@ for (const chain in networks) {
 
 const addresses = stagingAddresses.bridges
 const bonders = stagingAddresses.bonders
-const bonderFeeBps = stagingAddresses.bonderFeeBps
-const destinationFeeGasPriceMultiplier = stagingAddresses.destinationFeeGasPriceMultiplier
+const bonderFeeBps = stagingConfig.bonderFeeBps
+const destinationFeeGasPriceMultiplier = stagingConfig.destinationFeeGasPriceMultiplier
 
 export { addresses, chains, bonders, bonderFeeBps, destinationFeeGasPriceMultiplier }
